разделение методов
