+++ auth_lib/aiomethods/__init__.py
@@ -1,0 +1,4 @@
+from .auth_api import AsyncAuthAPILib
+from .userdata_api import AsyncUserdataAPILib
+
+__all__ = ["AsyncUserdataAPILib", "AsyncAuthAPILib"]
