+++ auth_lib/methods/__init__.py
@@ -1,0 +1,4 @@
+from .auth_api import AuthAPILib
+from .userdata_api import UserdataAPILib
+
+__all__ = ["AuthAPILib", "UserdataAPILib"]
